Improves interface for graphql parse schema
