+++ src/ParseServer.js
@@ -140,21 +140,11 @@ class ParseServer {
       req.config = req.config || Config.get(Parse.applicationId);
       req.auth = req.auth || new auth.Auth({ config: req.config });
       // TODO: only await perhaps once, and optimize perf
-      const schema = await Config.get(Parse.applicationId).database.loadSchema();
-      const allClasses = await schema.getAllClasses(true);
-      const classNames = [];
-      const fullSchema = allClasses.reduce((memo, classDef) => {
-        memo[classDef.className] = classDef;
-        classNames.push(classDef.className);
-        return memo;
-      }, {});
-      fullSchema.__classNames = classNames;
-      const Schema = new GraphQLParseSchema(Object.freeze(fullSchema));
-      const s = Schema.Schema();
-      const root = Schema.Root();
+      const graphQLSchema = new GraphQLParseSchema(Parse.applicationId);
+      const { schema, rootValue } = await graphQLSchema.load();
       return {
-        schema: s,
-        rootValue: root,
+        schema,
+        rootValue,
         graphiql
       }
     });
